Refactor python tests
Consolidate agent and server python3 tests.
