+++ lib/pbench/test/agent/conftest.py
@@ -69,7 +69,7 @@ def valid_config(pytestconfig):
 def invalid_config(pytestconfig):
     cfg_file = pytestconfig.cache.get("_PBENCH_AGENT_CONFIG", None)
     shutil.copyfile(
-        "./lib/pbench/test/unit/agent/config/pbench-agent.invalid.cfg", cfg_file
+        "./lib/pbench/test/agent/config/pbench-agent.invalid.cfg", cfg_file
     )
