+++ lib/pbench/test/agent/task/common.py
@@ -1,10 +1,10 @@
 import datetime
 
 FAKE_TIME = datetime.datetime(2020, 2, 2, 22, 0, 0)
-tarball = "lib/pbench/test/unit/agent/fixtures/copy_result_tb/log.tar.xz"
+tarball = "lib/pbench/test/agent/fixtures/copy_result_tb/log.tar.xz"
 bad_tarball = "nothing.tar.xz"
 
-MRT_DIR = "lib/pbench/test/unit/agent/fixtures/make_result_tb"
+MRT_DIR = "lib/pbench/test/agent/fixtures/make_result_tb"
 
 
 class MockDatetime(datetime.datetime):
@@ -14,4 +14,4 @@ class MockDatetime(datetime.datetime):
 
 
 def mock_agent_config(*args, **kwargs):
-    return {"pbench_run": "lib/pbench/test/unit/agent/fixtures/move_results"}
+    return {"pbench_run": "lib/pbench/test/agent/fixtures/move_results"}
