+++ lib/pbench/test/common/config/log-level.cfg
