+++ lib/pbench/test/common/config/logdir.cfg
