+++ lib/pbench/test/common/test_utils.py
@@ -8,7 +8,7 @@ class TestMd5sum:
     def test_md5sum():
         # Reuse the existing file from the server upload fixture.
         filename = "log.tar.xz"
-        test_file = Path("./lib/pbench/test/unit/server/fixtures/upload/", filename)
+        test_file = Path("./lib/pbench/test/server/fixtures/upload/", filename)
         expected_hash_md5 = open(f"{test_file}.md5", "r").read().split()[0]
         hash_md5 = md5sum(test_file)
         assert (
