+++ tox.ini
@@ -5,13 +5,12 @@ skip_missing_interpreters = True
 isolated_build = True
 toxworkdir = {env:TOXWORKDIR:/tmp/{env:USER}/tox}
 envlist = lint
-          py3-agent
+          py3
           datalog
           postprocess
           tool-scripts
           util-scripts
           bench-scripts
-          py3-server
           server
 
 [testenv]
@@ -81,25 +80,17 @@ deps =
 commands =
     bash -c "./agent/bench-scripts/unittests {posargs}"
 
-[testenv:py3-server]
-description = Runs all Python3-based server unit and functional tests
-deps =
-    -r{toxinidir}/server/requirements.txt
-    -r{toxinidir}/server/test-requirements.txt
-commands =
-    pytest ./lib/pbench/test/unit/common
-    pytest ./lib/pbench/test/unit/server 
-    pytest ./lib/pbench/test/functional
-
-[testenv:py3-agent]
-description = Runs all Python3-based agent unit and functional tests
+[testenv:py3]
+description = Runs all Python3-based tests
 deps =
     -r{toxinidir}/agent/requirements.txt
     -r{toxinidir}/agent/test-requirements.txt
+    -r{toxinidir}/server/requirements.txt
+    -r{toxinidir}/server/test-requirements.txt
 commands =
-    pytest ./lib/pbench/test/unit/common
-    pytest ./lib/pbench/test/unit/agent
-    pytest ./lib/pbench/test/functional
+    pytest ./lib/pbench/test/agent
+    pytest ./lib/pbench/test/server
+    pytest ./lib/pbench/test/common
 
 [testenv:lint]
 description = Runs all linting tasks
